requirements.txt
Flask==3.0.0
flask-cors==4.0.0
PyGithub==2.1.1
requests==2.31.0
gunicorn==21.2.0
